requirements.txt
streamlit
anthropic
pypdf
python-docx
